refactor(tools): migrate file tools to directory structure
Refactored File Operations (5 tools), Directory Operations (4 tools), and Path Utilities (8 tools) into organized directory structures following the established pattern. All 975 tests passing.
