+++ packages/tools/src/file/index.ts
@@ -1,10 +1,10 @@
 /**
  * File Tools
- * 
+ *
  * Tools for file system operations.
  */
 
-export * from './file-operations.js';
-export * from './directory-operations.js';
-export * from './path-utilities.js';
+export * from './operations/index.js';
+export * from './directory/index.js';
+export * from './path/index.js';
